refactor(core): split reranker.py under the 300-line cap (#199)
Behavior-preserving split of mcp_server/core/reranker.py (469 lines, 56%
over the repo's 300-line cap) into three cohesive core modules along the
seam the issue describes (scoring depends on lifecycle, never the reverse):

- Extract Class + Move Function -> reranker_model.py (135): model identity
  constants, durable cache_dir, offline-fetch gate (_offline_requested),
  RerankerStatus, _model_path, model_sha256. Pure, no singleton state.
- Extract Function -> reranker_scoring.py (168): _compute_retrieval_confidence,
  _compute_adaptive_alpha, _blend_scores. Pure ranking math, no I/O.
- reranker.py (238) keeps the process-global FlashRank singleton and the
  public rerank entrypoints (_ensure_reranker, ensure_reranker_loaded,
  reranker_status, rerank_results, get_raw_ce_score) and re-exports the
  companions so mcp_server.core.reranker.<name> stays stable.

The singleton state and _ensure_reranker stay in reranker.py because the
suite mutates reranker_mod._flashrank_instance and calls
reranker_mod._ensure_reranker() directly -- they must read/write one
module's globals. Every sourced comment and incident record (2026-07-10
silent-skip, /tmp cache_dir root cause, rejected Platt params + benchmark
numbers, adaptive-alpha ablation, 2026-07-27 timeout-less download) is
preserved, attached to the code it explains.

Before: reranker.py 469 lines (cap 300).
After:  reranker.py 238 / reranker_model.py 135 / reranker_scoring.py 168.
Tests: full suite 5361 passed / 21 failed / 95 skipped / 18 errors before
AND after (identical; the 21 fails + 18 errors are pre-existing psycopg-less
SQLite-env failures, none reranker-related, 0 test files edited).

Closes #199
